# lab8/openapi.yaml
openapi: 3.1.0
info:
  title: Library Async API with Rate Limiting
  version: 0.1.0
paths:
  /auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh Tokens
      operationId: refresh_tokens_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/:
    get:
      tags:
      - Books
      summary: Get Books
      operationId: get_books_books__get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 10
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'
                title: Response Get Books Books  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Books
      summary: Create Book
      operationId: create_book_books__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}:
    get:
      tags:
      - Books
      summary: Get Book
      operationId: get_book_books__book_id__get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Books
      summary: Delete Book
      operationId: delete_book_books__book_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BookCreate:
      properties:
        title:
          type: string
          minLength: 1
          title: Title
        author:
          type: string
          minLength: 1
          title: Author
        description:
          type: string
          title: Description
        status:
          $ref: '#/components/schemas/BookStatus'
          default: available
        year:
          type: integer
          exclusiveMinimum: 0.0
          title: Year
      type: object
      required:
      - title
      - author
      - description
      - year
      title: BookCreate
    BookResponse:
      properties:
        title:
          type: string
          minLength: 1
          title: Title
        author:
          type: string
          minLength: 1
          title: Author
        description:
          type: string
          title: Description
        status:
          $ref: '#/components/schemas/BookStatus'
          default: available
        year:
          type: integer
          exclusiveMinimum: 0.0
          title: Year
        _id:
          type: string
          title: Id
      type: object
      required:
      - title
      - author
      - description
      - year
      - _id
      title: BookResponse
    BookStatus:
      type: string
      enum:
      - available
      - issued
      title: BookStatus
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    RefreshTokenRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshTokenRequest
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: TokenResponse
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
          title: Username
        password:
          type: string
          maxLength: 50
          minLength: 6
          title: Password
      type: object
      required:
      - username
      - password
      title: UserCreate
    UserResponse:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 3
          title: Username
        _id:
          type: string
          title: Id
      type: object
      required:
      - username
      - _id
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer